moved flavor into its own module (within flavor module)
